fix: edit 페이지 카드 삭제시 명 수 변경
